Replace all the int id with long id

+++ Warehouse.Model/Branch.cs
@@ -9,7 +9,7 @@ namespace Warehouses.Model
 {
     public class Branch
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         [MaxLength(50)]
         [MinLength(3)]
         [Required]
@@ -22,6 +22,6 @@ namespace Warehouses.Model
         [MinLength(3)]
         [Required]
         public string Location { get; set; }
-        public int ParentId { get; set; }
+        public long ParentId { get; set; }
     }
 }
