+++ Warehouses.UI/Data/IBranchDataService.cs
@@ -6,9 +6,9 @@ namespace Warehouses.UI.Data
     public interface IBranchDataService
     {
         IEnumerable<Branch> GetAll();
-        IEnumerable<Branch> GetByParentId(int id);
+        IEnumerable<Branch> GetByParentId(long id);
         IEnumerable<Organization> GetAllOrganizations();
-        Branch GetById(int id);
+        Branch GetById(long id);
         bool HasSiblings();
         bool Delete(Branch model);
         bool Save(Branch model);
