+++ Warehouses.UI/Data/IOrganizationDataService.cs
@@ -6,7 +6,7 @@ namespace Warehouses.UI.Data
     public interface IOrganizationDataService
     {
         IEnumerable<Organization> GetAll();
-        Organization GetById(int id);
+        Organization GetById(long id);
         bool Save(Organization organization);
         bool Delete(Organization organization);
         bool HasSiblings();
