+++ Warehouses.UI/ViewModels/BranchDetailViewModel.cs
@@ -52,7 +52,7 @@ namespace Warehouses.UI.ViewModels
             }
         }
 
-        public override void Load(int id)
+        public override void Load(long id)
         {
             var branch = id > 0
               ? _branchService.GetById(id)
